# src/app/dashboard/docs/dashboard-api-openapi.yaml
openapi: 3.0.0
info:
  title: Badgr Dashboard API
  version: 1.0.0
  description: |
    Dashboard-specific API endpoints for the Badgr Server platform.

    These endpoints provide aggregated badge statistics, competency distribution,
    institution metrics, spatial data, and activity feeds optimized for dashboard
    visualizations.

  contact:
    name: ALGEBRA
    url: https://algebra-ec.de
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
security:
  - BearerAuth: []
...
paths:
  /v1/dashboard/badge-statistics:
    get:
      summary: Get aggregated badge statistics
      description: |
        Returns comprehensive badge statistics including total counts by type,
        trend analysis, and comparison to previous periods.

        **Use Case**: Dashboard KPI cards and summary metrics

        **Data Source**: badge_badgeclass and badge_badgeclassinstance tables

        **Performance**: Should complete in <100ms for typical users with <10k badges

      tags:
        - Badge Statistics
      operationId: getBadgeStatistics
      parameters:
        - name: include_trends
          in: query
          description: Include trend analysis comparing to previous period
          required: false
          schema:
            type: boolean
            default: true
        - name: period
          in: query
          description: Period for trend comparison (30d, 90d, 1y)
          required: false
          schema:
            type: string
            enum: [30d, 90d, 1y]
            default: 30d
      responses:
        '200':
          description: Successful response with badge statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadgeStatisticsResponse'
              example:
                total: 1247
                byType:
                  competency: 523
                  participation: 489
                  learningpath: 235
                trend:
                  direction: up
                  percentageChange: 12.5
                  absoluteChange: 138
                  comparisonPeriod: 30d
                  previousTotal: 1109
                lastUpdated: '2025-10-26T13:00:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/dashboard/institution-statistics:
    get:
      summary: Get institution-level statistics
      description: |
        Returns aggregated statistics about badge-issuing institutions including
        total counts, active institutions, new registrations, and regional distribution.

        **Use Case**: Institution KPIs, organizational dashboard cards

        **Data Source**: issuer_issuer table with aggregations

        **Admin Access**: May require elevated permissions for full statistics

      tags:
        - Institution Statistics
      operationId: getInstitutionStatistics
      parameters:
        - name: include_regional
          in: query
          description: Include breakdown by region/canton
          required: false
          schema:
            type: boolean
            default: true
        - name: time_period
          in: query
          description: Period for "new institutions" calculation
          required: false
          schema:
            type: string
            enum: [30d, 90d, 1y]
            default: 30d
      responses:
        '200':
          description: Successful response with institution statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionStatisticsResponse'
              example:
                total: 89
                active: 67
                new: 5
                byRegion:
                  Zürich: 23
                  Bern: 18
                  Basel: 12
                  Luzern: 9
                  Geneva: 8
                  other: 19
                byType:
                  university: 15
                  vocational_school: 34
                  company: 28
                  nonprofit: 12
                averageBadgesPerInstitution: 14.2
                topIssuers:
                  - name: ETH Zürich
                    slug: eth-zurich
                    badgeCount: 127
                  - name: Swisscom
                    slug: swisscom
                    badgeCount: 98
                lastUpdated: '2025-10-26T13:00:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/dashboard/activity-feed:
    get:
      summary: Get recent badge and activity events
      description: |
        Returns a paginated feed of recent platform activities including
        badge awards, new institutions, collection creations, and skill additions.

        **Use Case**: Activity timeline widgets, notification feeds

        **Data Source**: Multiple tables with timestamp-based queries

        **Real-time**: Consider WebSocket/SSE for live updates in future

      tags:
        - Activity Feed
      operationId: getActivityFeed
      parameters:
        - name: limit
          in: query
          description: Number of activity items to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of items to skip (for pagination)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: types
          in: query
          description: Filter by activity types (comma-separated)
          required: false
          schema:
            type: string
            example: badge_issued,collection_created
        - name: since
          in: query
          description: Return only activities after this timestamp
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response with activity feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityFeedResponse'
              example:
                results:
                  - id: act-12345
                    type: badge_issued
                    title: Neues Digitalkompetenz Badge verliehen
                    description: Digital Leadership Badge an 5 Personen
                    count: 5
                    timestamp: '2025-10-26T12:45:00Z'
                    metadata:
                      badgeName: Digital Leadership
                      issuerName: ETH Zürich
                    icon: award
                  - id: act-12344
                    type: institution_added
                    title: Neue Institution registriert
                    description: Swisscom AG beigetreten
                    count: 1
                    timestamp: '2025-10-26T11:20:00Z'
                    metadata:
                      institutionName: Swisscom AG
                    icon: building
                  - id: act-12343
                    type: collection_created
                    title: Badge-Sammlung erstellt
                    description: 'Neue Sammlung: IT Security Basics'
                    count: 1
                    timestamp: '2025-10-26T10:15:00Z'
                    metadata:
                      collectionName: IT Security Basics
                      badgeCount: 8
                    icon: folder
                count: 247
                next: /v1/dashboard/activity-feed?offset=10&limit=10
                previous: null
                lastUpdated: '2025-10-26T13:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /o/token or /v1/user/auth-token endpoints.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    TrendDirection:
      type: string
      enum:
        - up
        - down
        - stable
      description: |
        Trend direction indicator:
        - up: Increasing trend (>2% change)
        - down: Decreasing trend (<-2% change)
        - stable: No significant change (-2% to +2%)
    BadgeStatisticsResponse:
      type: object
      required:
        - total
        - byType
        - lastUpdated
      properties:
        total:
          type: integer
          description: Total number of badges across all types
          example: 1247
        byType:
          type: object
          required:
            - competency
            - participation
            - learningpath
          properties:
            competency:
              type: integer
              description: Number of competency badges
              example: 523
            participation:
              type: integer
              description: Number of participation badges
              example: 489
            learningpath:
              type: integer
              description: Number of learning path badges
              example: 235
        trend:
          type: object
          description: Trend analysis compared to previous period
          properties:
            direction:
              $ref: '#/components/schemas/TrendDirection'
            percentageChange:
              type: number
              format: float
              description: Percentage change from previous period
              example: 12.5
            absoluteChange:
              type: integer
              description: Absolute number change from previous period
              example: 138
            comparisonPeriod:
              type: string
              description: Period used for comparison
              example: 30d
            previousTotal:
              type: integer
              description: Total from previous period
              example: 1109
        lastUpdated:
          type: string
          format: date-time
          description: Timestamp of last data update
          example: '2025-10-26T13:00:00Z'
...
    TopIssuer:
      type: object
      required:
        - name
        - slug
        - badgeCount
      properties:
        name:
          type: string
          description: Institution display name
          example: ETH Zürich
        slug:
          type: string
          description: Institution URL slug identifier
          example: eth-zurich
        badgeCount:
          type: integer
          description: Number of badges issued
          example: 127
    InstitutionStatisticsResponse:
      type: object
      required:
        - total
        - active
        - lastUpdated
      properties:
        total:
          type: integer
          description: Total number of registered institutions
          example: 89
        active:
          type: integer
          description: Number of active institutions (issued badges recently)
          example: 67
        new:
          type: integer
          description: Number of newly registered institutions in time period
          example: 5
        byRegion:
          type: object
          description: Institution count by region/canton
          additionalProperties:
            type: integer
          example:
            Zürich: 23
            Bern: 18
            Basel: 12
        byType:
          type: object
          description: Institution count by organization type
          additionalProperties:
            type: integer
          example:
            university: 15
            vocational_school: 34
            company: 28
        averageBadgesPerInstitution:
          type: number
          format: float
          description: Average badge classes per institution
          example: 14.2
        topIssuers:
          type: array
          description: Top badge-issuing institutions
          items:
            $ref: '#/components/schemas/TopIssuer'
        lastUpdated:
          type: string
          format: date-time
          example: '2025-10-26T13:00:00Z'
...
    ActivityType:
      type: string
      enum:
        - badge_issued
        - collection_created
        - institution_added
        - skill_added
        - badge_created
      description: Type of platform activity
    ActivityItem:
      type: object
      required:
        - id
        - type
        - title
        - timestamp
      properties:
        id:
          type: string
          description: Unique activity identifier
          example: act-12345
        type:
          $ref: '#/components/schemas/ActivityType'
        title:
          type: string
          description: Activity title (localized)
          example: Neues Digitalkompetenz Badge verliehen
        description:
          type: string
          description: Activity description
          example: Digital Leadership Badge an 5 Personen
        count:
          type: integer
          description: Number of items affected
          example: 5
        timestamp:
          type: string
          format: date-time
          description: Activity timestamp
          example: '2025-10-26T12:45:00Z'
        metadata:
          type: object
          description: Additional activity metadata
          additionalProperties: true
          example:
            badgeName: Digital Leadership
            issuerName: ETH Zürich
        icon:
          type: string
          description: Icon identifier for UI display
          example: award
    ActivityFeedResponse:
      type: object
      required:
        - results
        - count
        - lastUpdated
      properties:
        results:
          type: array
          description: Array of activity items
          items:
            $ref: '#/components/schemas/ActivityItem'
        count:
          type: integer
          description: Total number of activities matching query
          example: 247
        next:
          type: string
          format: uri
          description: URL for next page of results
          example: /v1/dashboard/activity-feed?offset=10&limit=10
          nullable: true
        previous:
          type: string
          format: uri
          description: URL for previous page of results
          nullable: true
        lastUpdated:
          type: string
          format: date-time
          example: '2025-10-26T13:00:00Z'
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code or type
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid date format for startDate parameter
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: '2025-10-26T13:00:00Z'
  responses:
    BadRequestError:
      description: Bad request - Invalid parameters or request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BAD_REQUEST
            message: Invalid date format for startDate parameter
            details:
              parameter: startDate
              expected: ISO 8601 date format (YYYY-MM-DD)
              received: '2024-13-45'
            timestamp: '2025-10-26T13:00:00Z'
    UnauthorizedError:
      description: Unauthorized - Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Authentication required. Please provide a valid Bearer token.
            timestamp: '2025-10-26T13:00:00Z'
    ForbiddenError:
      description: Forbidden - Insufficient permissions for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: FORBIDDEN
            message: Insufficient permissions to access institution statistics
            details:
              required_permission: view_institution_stats
              user_role: earner
            timestamp: '2025-10-26T13:00:00Z'
...
    InternalServerError:
      description: Internal server error - Unexpected server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: INTERNAL_SERVER_ERROR
            message: An unexpected error occurred while processing your request
            timestamp: '2025-10-26T13:00:00Z'